feat(gpio_driver): Make GPIO more compatible with F4xx line
GPIO banks are now checked for existence before usage. The cost of this is a big macro mess.
Why a serial switch case instead of a boundary check like in ADC? Well because ST has GPIO banks
that go like this: A, B, C, H. Boundary check of the letter H would allow intermediate values as
well and it's hard to go around it.
Applied necessary changes in unit tests.